Add iOS support

+++ hooks/before_prepare/java_copy.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
-cp android/MainActivity.java ./platforms/android/app/src/main/java/io/cordova/hellocordova/MainActivity.java
+[ -e ./platforms/android ] && cp android/MainActivity.java ./platforms/android/app/src/main/java/io/cordova/hellocordova/MainActivity.java
+exit 0
